\name{data1}
\alias{data1}
\docType{data}
\title{
data1
}
\description{
A time series with one column.
}
\usage{data("data1")}
\format{
  The format is:
 Time-Series [1:200] from 0 to 1: -0.089595464  0.017667207 -0.045008295 -0.115746279 -0.117493959 ...
}
\details{
A single vector with one column:
}

\examples{
library(DFA)
data("data1")
}
\keyword{datasets}
