\name{data2}
\alias{data2}
\docType{data}
\title{
data2
}
\description{
A time series with one column.
}
\usage{data("data2")}
\format{
  The format is:
 Time-Series [1:200] from 0 to 1:  0.0293948 -0.0315427 -0.1081185 -0.2326447 -0.2165815 ...
}
\details{
A single vector with one column:
}

\examples{
library(DFA)
data("data2")
}
\keyword{datasets}
