\name{data3}
\alias{data3}
\docType{data}
\title{
data3
}
\description{
A time series with one column.
}
\usage{data("data3")}
\format{
  The format is:
 Time-Series [1:200] from 0 to 1: -0.04827469 -0.02439724 -0.09959182 -0.16912343 -0.18852878  ...
}
\details{
A single vector with one column:
}
\examples{
library(DFA)
data("data3")
}
\keyword{datasets}
