\name{data4}
\alias{data4}
\docType{data}
\title{
data4}
\description{
A time series with one column.
}
\usage{data("data4")}
\format{
  The format is:
 Time-Series [1:200] from 0 to 1: -0.019939448 -0.097243289 -0.091787349 -0.044202712  0.037363582 ...
}
\details{
A single vector with one column:
}
\examples{
library(DFA)
data("data4")

}
\keyword{datasets}
