% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Master_functions.R
\name{DFA_pred}
\alias{DFA_pred}
\title{A function to predict groups assignments.}
\usage{
DFA_pred(model, data)
}
\arguments{
\item{model}{groups ~ predictor variables}

\item{data}{data used to predict group assignments}
}
\description{
This function takes a specified model of your groups ~ predictor variables and then uses that model to predict group membership of you specified data.
}
\keyword{Prediction}
