% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDFA-package.R
\docType{package}
\name{LDFA-package}
\alias{LDFA}
\alias{LDFA-package}
\title{LDFA: DFA Package}
\description{
This package allows you to run a simple linear discriminant function analysis.
}
\author{
\strong{Maintainer}: David, Hans Ledesma, Bilger \email{ledesma-david@utexas.edu}

}
\keyword{internal}
